Add custom hook for reduced motion preference. Create src/data/faqs.json
